backend/requirements.txt
fastapi
uvicorn
pydantic
groq
requests
python-dotenv
pdfplumber
pandas
openpyxl
pytesseract
Pillow
python-multipart
